% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASTER-CSM-APPLICATION.R
\name{APPLICATION_CSM_MOY_GLOBAL}
\alias{APPLICATION_CSM_MOY_GLOBAL}
\title{CENTRALIZED STATISTICAL MONITORING : MASTER PROGRAM FOR CSM APPLICATION}
\usage{
APPLICATION_CSM_MOY_GLOBAL(data = data)
}
\arguments{
\item{data}{is the table containing a multicenter clinical trial data on which the function program for each CSM methode will be performed.}
}
\value{
table
}
\description{
CENTRALIZED STATISTICAL MONITORING : MASTER PROGRAM FOR CSM APPLICATION
}
