% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSM-Bayesian-HATAYAMA.R
\name{CONSTRUCT_BAYES}
\alias{CONSTRUCT_BAYES}
\title{CENTRALIZED STATISTICAL MONITORING : HATAYAMA METHOD
Function to implemente the Hatayama CSM method which uses the Bayesian Finite
Mixture Models (FMM) in order to detect an atypical center in multicenter trial.
This method like the others of this package implementes the CSM for the
 continuous variable data}
\usage{
CONSTRUCT_BAYES(dataHatayama)
}
\arguments{
\item{dataHatayama}{}
}
\value{
Table
}
\description{
CENTRALIZED STATISTICAL MONITORING : HATAYAMA METHOD
Function to implemente the Hatayama CSM method which uses the Bayesian Finite
Mixture Models (FMM) in order to detect an atypical center in multicenter trial.
This method like the others of this package implementes the CSM for the
 continuous variable data
}
