% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSM-Distance.R
\name{CONSTRUCT_DISTANCE}
\alias{CONSTRUCT_DISTANCE}
\title{CENTRALIZED STATISTICAL MONITORING : DISTANCE METHOD}
\usage{
CONSTRUCT_DISTANCE(data = data)
}
\arguments{
\item{data}{a multicenter clinical trial data containing one or several continuous variable}
}
\value{
table
}
\description{
CENTRALIZED STATISTICAL MONITORING : DISTANCE METHOD
}
