% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSM-Student-Test.R
\name{CONSTRUCT_TestMOY}
\alias{CONSTRUCT_TestMOY}
\title{CENTRALIZED STATISTICAL MONITORING : STUDENT TEST METHOD}
\usage{
CONSTRUCT_TestMOY(data = data)
}
\arguments{
\item{data}{a multicenter clinical trial data containing one or several continuous variable}
}
\value{
table
}
\description{
CENTRALIZED STATISTICAL MONITORING : STUDENT TEST METHOD
}
