% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSM-MixedLinear-DESMET.R
\name{CONSTRUCT_ZDI}
\alias{CONSTRUCT_ZDI}
\title{CENTRALIZED STATISTICAL MONITORING : DESMET METHOD}
\usage{
CONSTRUCT_ZDI(data = data)
}
\arguments{
\item{data}{a multicenter clinical trial data containing one or several continuous variable}
}
\value{
table
}
\description{
CENTRALIZED STATISTICAL MONITORING : DESMET METHOD
}
