% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSM-Categorical-DESMET.R
\name{CSM_BETA_BINOMIAL_SIMS}
\alias{CSM_BETA_BINOMIAL_SIMS}
\title{CENTRALIZED STATISTICAL MONITORING : DESMET CATEGORICAL CSM METHOD USING BETA-BINOMIAL MODELING
Function to implemente the Desmet Categorical CSM method which uses the Betabinomial modelling
in order to detect an atypical center in multicenter trial for a categorical binary variable.
This method like the others of this package implementes the CSM for the
 categorical variable data}
\usage{
CSM_BETA_BINOMIAL_SIMS(bdata = bdata)
}
\arguments{
\item{dataHatayama}{}
}
\value{
Table
}
\description{
CENTRALIZED STATISTICAL MONITORING : DESMET CATEGORICAL CSM METHOD USING BETA-BINOMIAL MODELING
Function to implemente the Desmet Categorical CSM method which uses the Betabinomial modelling
in order to detect an atypical center in multicenter trial for a categorical binary variable.
This method like the others of this package implementes the CSM for the
 categorical variable data
}
