% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGenerate.R
\name{data.generebis}
\alias{data.generebis}
\title{MONITORING STATISTIQUE DES ESSAIS : TRAVAUX DE SIMULATION}
\usage{
data.generebis(
  n.subject = 50,
  n.center = 6,
  sigma2c = 0.7,
  sigma2r = 0.5,
  mu = 100,
  centreAtypic = 2,
  delta = 10
)
}
\arguments{
\item{n.subject}{is the number of subject in a site (considering that all centers have the same subjects number)}

\item{n.center}{indicate the number of investigative sites in the multicenter trials. His default value \code{6}}

\item{sigma2c}{description}

\item{sigma2r}{description}

\item{mu}{description}

\item{centreAtypic}{description}

\item{delta}{description}
}
\value{
Table
}
\description{
MONITORING STATISTIQUE CENTRALISE (CSM) : DATA GENERATION FOR CONTINUOUS VARIABLE
}
