% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASTER-CSM-SIMULATION.R
\name{MASTER_CSM_MOY_GLOBAL_SIMS}
\alias{MASTER_CSM_MOY_GLOBAL_SIMS}
\title{CENTRALIZED STATISTICAL MONITORING : MASTER PROGRAM FOR CSM SIMULATION}
\usage{
MASTER_CSM_MOY_GLOBAL_SIMS(
  NSiteAtypiq = 1,
  MaxEcart = 2,
  NombreSimul = 2,
  n.subject = 50,
  n.center = 6,
  sigma2c = 1,
  sigma2r = 4,
  mu = 0
)
}
\arguments{
\item{MaxEcart}{. His default value \code{6}}

\item{NombreSimul}{a number of replications of simulations. His default value is \code{5}}

\item{n.subject}{is the number of subject in a site (considering that all centers have the same subjects number)}

\item{n.center}{indicate the number of investigative sites in the multicenter trials. His default value \code{6}}

\item{sigma2c}{. His default value \code{1}}

\item{sigma2r}{. His default value \code{4} according the default value of the mean}

\item{mu}{is the mean of the non atypical centers. His default value \code{10}}

\item{SiteAtypiq}{is the maximum number of atypical site in the simulations, that is the number of atypical site will varying from 1 to SiteAtypiq. His default value \code{6}}
}
\value{
list of tables that contains in one the performance (ie sensitivity and specificity) of methods, and in the other the assignation data
}
\description{
CENTRALIZED STATISTICAL MONITORING : MASTER PROGRAM FOR CSM SIMULATION
}
